change CellViewModel - move to own file
