+++ VKApp/VKApp/Models/CellItemsMock/UserCellModelFactory.swift
@@ -5,26 +5,9 @@
 //  Created by Artem Mayer on 11.01.2023.
 //
 
-import SwiftUI
+import Foundation
 
-// MARK: - CellViewModel
-
-struct CellViewModel: Identifiable {
-    let id: Int
-    let fullName: String
-    let avatar: URL?
-}
-
-// MARK: - GrouppedUserModel
-
-struct GrouppedUserModel: Identifiable {
-    var id: Character
-    var users: [CellViewModel]
-}
-
-// MARK: - UserCellModelFactory
-
-final class UserCellModelFactory {
+struct UserCellModelFactory {
 
     // MARK: - Functions
